+++ smassh/src/css_generator.py
@@ -1,4 +1,4 @@
-from appdirs import user_cache_dir
+from platformdirs import user_cache_dir
 from pathlib import Path
 
 TARGET_FOLDER = Path(user_cache_dir("smassh"))